requirements.txt
flask==3.0.2
flask-cors==4.0.0
openai==1.12.0
tiktoken
python-dotenv==1.0.1
psycopg2-binary
httpx==0.27.2
